requirement.txt
boto3
fastapi
pymysql
python-dotenv
python-multipart
uvicorn[standard]